+++ output/pokedex.csv
@@ -1,0 +1,152 @@
+#,Name,Type,Total,HP,Attack,Defense,Sp. Atk,Sp. Def,Speed
+001,Bulbasaur,GRASS/POISON,318,45,49,49,65,65,45
+002,Ivysaur,GRASS/POISON,405,60,62,63,80,80,60
+003,Venusaur,GRASS/POISON,525,80,82,83,100,100,80
+004,Charmander,FIRE,309,39,52,43,60,50,65
+005,Charmeleon,FIRE,405,58,64,58,80,65,80
+006,Charizard,FIRE/FLYING,534,78,84,78,109,85,100
+007,Squirtle,WATER,314,44,48,65,50,64,43
+008,Wartortle,WATER,405,59,63,80,65,80,58
+009,Blastoise,WATER,530,79,83,100,85,105,78
+010,Caterpie,BUG,195,45,30,35,20,20,45
+011,Metapod,BUG,205,50,20,55,25,25,30
+012,Butterfree,BUG/FLYING,395,60,45,50,90,80,70
+013,Weedle,BUG/POISON,195,40,35,30,20,20,50
+014,Kakuna,BUG/POISON,205,45,25,50,25,25,35
+015,Beedrill,BUG/POISON,395,65,90,40,45,80,75
+016,Pidgey,NORMAL/FLYING,251,40,45,40,35,35,56
+017,Pidgeotto,NORMAL/FLYING,349,63,60,55,50,50,71
+018,Pidgeot,NORMAL/FLYING,479,83,80,75,70,70,101
+019,Rattata,NORMAL,253,30,56,35,25,35,72
+020,Raticate,NORMAL,413,55,81,60,50,70,97
+021,Spearow,NORMAL/FLYING,262,40,60,30,31,31,70
+022,Fearow,NORMAL/FLYING,442,65,90,65,61,61,100
+023,Ekans,POISON,288,35,60,44,40,54,55
+024,Arbok,POISON,448,60,95,69,65,79,80
+025,Pikachu,ELECTRIC,320,35,55,40,50,50,90
+026,Raichu,ELECTRIC,485,60,90,55,90,80,110
+027,Sandshrew,GROUND,300,50,75,85,20,30,40
+028,Sandslash,GROUND,450,75,100,110,45,55,65
+029,Nidoran♀,POISON,275,55,47,52,40,40,41
+030,Nidorina,POISON,365,70,62,67,55,55,56
+031,Nidoqueen,POISON/GROUND,505,90,92,87,75,85,76
+032,Nidoran♂,POISON,273,46,57,40,40,40,50
+033,Nidorino,POISON,365,61,72,57,55,55,65
+034,Nidoking,POISON/GROUND,505,81,102,77,85,75,85
+035,Clefairy,FAIRY,323,70,45,48,60,65,35
+036,Clefable,FAIRY,483,95,70,73,95,90,60
+037,Vulpix,FIRE,299,38,41,40,50,65,65
+038,Ninetales,FIRE,505,73,76,75,81,100,100
+039,Jigglypuff,NORMAL/FAIRY,270,115,45,20,45,25,20
+040,Wigglytuff,NORMAL/FAIRY,435,140,70,45,85,50,45
+041,Zubat,POISON/FLYING,245,40,45,35,30,40,55
+042,Golbat,POISON/FLYING,455,75,80,70,65,75,90
+043,Oddish,GRASS/POISON,320,45,50,55,75,65,30
+044,Gloom,GRASS/POISON,395,60,65,70,85,75,40
+045,Vileplume,GRASS/POISON,490,75,80,85,110,90,50
+046,Paras,BUG/GRASS,285,35,70,55,45,55,25
+047,Parasect,BUG/GRASS,405,60,95,80,60,80,30
+048,Venonat,BUG/POISON,305,60,55,50,40,55,45
+049,Venomoth,BUG/POISON,450,70,65,60,90,75,90
+050,Diglett,GROUND,265,10,55,25,35,45,95
+051,Dugtrio,GROUND,425,35,100,50,50,70,120
+052,Meowth,NORMAL,290,40,45,35,40,40,90
+053,Persian,NORMAL,440,65,70,60,65,65,115
+054,Psyduck,WATER,320,50,52,48,65,50,55
+055,Golduck,WATER,500,80,82,78,95,80,85
+056,Mankey,FIGHTING,305,40,80,35,35,45,70
+057,Primeape,FIGHTING,455,65,105,60,60,70,95
+058,Growlithe,FIRE,350,55,70,45,70,50,60
+059,Arcanine,FIRE,555,90,110,80,100,80,95
+060,Poliwag,WATER,300,40,50,40,40,40,90
+061,Poliwhirl,WATER,385,65,65,65,50,50,90
+062,Poliwrath,WATER/FIGHTING,510,90,95,95,70,90,70
+063,Abra,PSYCHIC,310,25,20,15,105,55,90
+064,Kadabra,PSYCHIC,400,40,35,30,120,70,105
+065,Alakazam,PSYCHIC,500,55,50,45,135,95,120
+066,Machop,FIGHTING,305,70,80,50,35,35,35
+067,Machoke,FIGHTING,405,80,100,70,50,60,45
+068,Machamp,FIGHTING,505,90,130,80,65,85,55
+069,Bellsprout,GRASS/POISON,300,50,75,35,70,30,40
+070,Weepinbell,GRASS/POISON,390,65,90,50,85,45,55
+071,Victreebel,GRASS/POISON,490,80,105,65,100,70,70
+072,Tentacool,WATER/POISON,335,40,40,35,50,100,70
+073,Tentacruel,WATER/POISON,515,80,70,65,80,120,100
+074,Geodude,ROCK/GROUND,300,40,80,100,30,30,20
+075,Graveler,ROCK/GROUND,390,55,95,115,45,45,35
+076,Golem,ROCK/GROUND,495,80,120,130,55,65,45
+077,Ponyta,FIRE,410,50,85,55,65,65,90
+078,Rapidash,FIRE,500,65,100,70,80,80,105
+079,Slowpoke,WATER/PSYCHIC,315,90,65,65,40,40,15
+080,Slowbro,WATER/PSYCHIC,490,95,75,110,100,80,30
+081,Magnemite,ELECTRIC/STEEL,325,25,35,70,95,55,45
+082,Magneton,ELECTRIC/STEEL,465,50,60,95,120,70,70
+083,Farfetch'd,NORMAL/FLYING,377,52,90,55,58,62,60
+084,Doduo,NORMAL/FLYING,310,35,85,45,35,35,75
+085,Dodrio,NORMAL/FLYING,470,60,110,70,60,60,110
+086,Seel,WATER,325,65,45,55,45,70,45
+087,Dewgong,WATER/ICE,475,90,70,80,70,95,70
+088,Grimer,POISON,325,80,80,50,40,50,25
+089,Muk,POISON,500,105,105,75,65,100,50
+090,Shellder,WATER,305,30,65,100,45,25,40
+091,Cloyster,WATER/ICE,525,50,95,180,85,45,70
+092,Gastly,GHOST/POISON,310,30,35,30,100,35,80
+093,Haunter,GHOST/POISON,405,45,50,45,115,55,95
+094,Gengar,GHOST/POISON,500,60,65,60,130,75,110
+095,Onix,ROCK/GROUND,385,35,45,160,30,45,70
+096,Drowzee,PSYCHIC,328,60,48,45,43,90,42
+097,Hypno,PSYCHIC,483,85,73,70,73,115,67
+098,Krabby,WATER,325,30,105,90,25,25,50
+099,Kingler,WATER,475,55,130,115,50,50,75
+100,Voltorb,ELECTRIC,330,40,30,50,55,55,100
+101,Electrode,ELECTRIC,490,60,50,70,80,80,150
+102,Exeggcute,GRASS/PSYCHIC,325,60,40,80,60,45,40
+103,Exeggutor,GRASS/PSYCHIC,530,95,95,85,125,75,55
+104,Cubone,GROUND,320,50,50,95,40,50,35
+105,Marowak,GROUND,425,60,80,110,50,80,45
+106,Hitmonlee,FIGHTING,455,50,120,53,35,110,87
+107,Hitmonchan,FIGHTING,455,50,105,79,35,110,76
+108,Lickitung,NORMAL,385,90,55,75,60,75,30
+109,Koffing,POISON,340,40,65,95,60,45,35
+110,Weezing,POISON,490,65,90,120,85,70,60
+111,Rhyhorn,GROUND/ROCK,345,80,85,95,30,30,25
+112,Rhydon,GROUND/ROCK,485,105,130,120,45,45,40
+113,Chansey,NORMAL,450,250,5,5,35,105,50
+114,Tangela,GRASS,435,65,55,115,100,40,60
+115,Kangaskhan,NORMAL,490,105,95,80,40,80,90
+116,Horsea,WATER,295,30,40,70,70,25,60
+117,Seadra,WATER,440,55,65,95,95,45,85
+118,Goldeen,WATER,320,45,67,60,35,50,63
+119,Seaking,WATER,450,80,92,65,65,80,68
+120,Staryu,WATER,340,30,45,55,70,55,85
+121,Starmie,WATER/PSYCHIC,520,60,75,85,100,85,115
+122,Mr. Mime,PSYCHIC/FAIRY,460,40,45,65,100,120,90
+123,Scyther,BUG/FLYING,500,70,110,80,55,80,105
+124,Jynx,ICE/PSYCHIC,455,65,50,35,115,95,95
+125,Electabuzz,ELECTRIC,490,65,83,57,95,85,105
+126,Magmar,FIRE,495,65,95,57,100,85,93
+127,Pinsir,BUG,500,65,125,100,55,70,85
+128,Tauros,NORMAL,490,75,100,95,40,70,110
+129,Magikarp,WATER,200,20,10,55,15,20,80
+130,Gyarados,WATER/FLYING,540,95,125,79,60,100,81
+131,Lapras,WATER/ICE,535,130,85,80,85,95,60
+132,Ditto,NORMAL,288,48,48,48,48,48,48
+133,Eevee,NORMAL,325,55,55,50,45,65,55
+134,Vaporeon,WATER,525,130,65,60,110,95,65
+135,Jolteon,ELECTRIC,525,65,65,60,110,95,130
+136,Flareon,FIRE,525,65,130,60,95,110,65
+137,Porygon,NORMAL,395,65,60,70,85,75,40
+138,Omanyte,ROCK/WATER,355,35,40,100,90,55,35
+139,Omastar,ROCK/WATER,495,70,60,125,115,70,55
+140,Kabuto,ROCK/WATER,355,30,80,90,55,45,55
+141,Kabutops,ROCK/WATER,495,60,115,105,65,70,80
+142,Aerodactyl,ROCK/FLYING,515,80,105,65,60,75,130
+143,Snorlax,NORMAL,540,160,110,65,65,110,30
+144,Articuno,ICE/FLYING,580,90,85,100,95,125,85
+145,Zapdos,ELECTRIC/FLYING,580,90,90,85,125,90,100
+146,Moltres,FIRE/FLYING,580,90,100,90,125,85,90
+147,Dratini,DRAGON,300,41,64,45,50,50,50
+148,Dragonair,DRAGON,420,61,84,65,70,70,70
+149,Dragonite,DRAGON/FLYING,600,91,134,95,100,100,80
+150,Mewtwo,PSYCHIC,680,106,110,90,154,90,130
+151,Mew,PSYCHIC,600,100,100,100,100,100,100
